% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmML.R
\name{bmMLE}
\alias{bmMLE}
\title{Brownian motion maximum likelihood estimation}
\usage{
bmMLE(phy, x)
}
\arguments{
\item{phy}{a phylogeny object of class \code{phylo} from package \code{ape}}

\item{x}{a numeric vector of traits, length equal to number of tips in
\code{phy}; order of trait values is assumed (and not checked) to be same as
order of tips in \code{phy}}
}
\value{
a named list giving the MLE of the root trait (\code{x0}), BM
variance (\code{s2}), and variance-covariance matrix (\code{Smat})
}
\description{
Fits the parameters of a Brownian motion model to phylogenetic
and trait data
}
\details{
stub
}
